migrate external event handlers from http.py
